# public/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Event Schedule API",
    "version": "1.0",
    "description": "API for managing events, schedules, and tickets on Event Schedule. To get started, register an account, then create a schedule before adding events. Write operations require a Pro plan. New accounts get a trial period."
  },
  "servers": [
    {
      "url": "{baseUrl}/api",
      "variables": {
        "baseUrl": {
          "default": "https://eventschedule.com"
        }
      }
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/register": {
      "post": {
        "operationId": "register",
        "tags": ["Authentication"],
        "summary": "Register a new account",
        "description": "Creates a new user account and returns an API key. In hosted mode, a verification code (from /register/send-code) is required. In selfhosted mode, registration is only allowed if no users exist yet. Rate limited to 3 registrations per IP per hour.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 255,
                    "description": "Full name of the user"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255,
                    "description": "Email address for the account"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "description": "Account password (minimum 8 characters)"
                  },
                  "verification_code": {
                    "type": "string",
                    "description": "6-digit verification code from /register/send-code (required in hosted mode only)"
                  },
                  "timezone": {
                    "type": "string",
                    "maxLength": 100,
                    "description": "IANA timezone identifier",
                    "example": "America/New_York"
                  },
                  "language_code": {
                    "type": "string",
                    "description": "Preferred language code",
                    "enum": ["en", "es", "de", "fr", "it", "pt", "he", "nl", "ar", "et", "ru"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "api_key": {
                          "type": "string",
                          "description": "API key for authenticating future requests. Store this securely."
                        },
                        "api_key_expires_at": {
                          "type": "string",
                          "format": "date-time",
                          "description": "ISO 8601 timestamp when the API key expires"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "description": "Encoded user ID"
                            },
                            "name": {
                              "type": "string",
                              "description": "User's full name"
                            },
                            "email": {
                              "type": "string",
                              "format": "email",
                              "description": "User's email address"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Registration is closed (selfhosted mode with existing users)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too many registration attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "operationId": "login",
        "tags": ["Authentication"],
        "summary": "Login to an existing account",
        "description": "Authenticates with email and password, returning a new API key. If an active API key already exists, returns a 409 error - disable and re-enable API access in account settings to generate a new key. Store the key securely. Accounts with two-factor authentication enabled must generate an API key from the web UI instead. Rate limited to 5 attempts per IP per 15 minutes.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Account email address"
                  },
                  "password": {
                    "type": "string",
                    "description": "Account password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "api_key": {
                          "type": "string",
                          "description": "New API key for authenticating future requests. Store this securely."
                        },
                        "api_key_expires_at": {
                          "type": "string",
                          "format": "date-time",
                          "description": "ISO 8601 timestamp when the API key expires"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "description": "Encoded user ID"
                            },
                            "name": {
                              "type": "string",
                              "description": "User's full name"
                            },
                            "email": {
                              "type": "string",
                              "format": "email",
                              "description": "User's email address"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "An active API key already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Account has two-factor authentication enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too many login attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key obtained from /login or /register. Keys expire after one year."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "object",
            "nullable": true,
            "description": "Field-level validation errors (keyed by field name, values are arrays of error messages)",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": ["error"]
      }
    }
  },
...
}